Implement parsing the block structure using an octree
The octree describes the structure of the mesh. The parent is the big block and the children are smaller (1/8 of parents size) blocks and are in the parent block.

The height of a node describes it refinement level. The higher the level, the smaller the block.

Hint: There is NO data attached to the commit for tests! You must modify the path of the tests at the moment to pass them.